+++ src/ItreeNet/Data/Models/ProvisorischeBuchung.cs
@@ -4,7 +4,7 @@ namespace ItreeNet.Data.Models
     {
         public Guid Id { get; set; }
         public DateOnly Datum { get; set; }
-        public decimal? Zeit { get; set; }
+        public int? Zeit { get; set; }
         public string Buchungstext { get; set; } = string.Empty;
         public string? ProjektName { get; set; }
         public string? VorgangName { get; set; }
